ha i'm a dumb
relooking at my code, it looks like thread delay time is what i need, as
afterwards, when it runs the loop again, it'll check buttons again.
but if the button is pressed DURING THE DELAY, and the button isn't
being pressed afterwards, it'll read that no button is pressed.

TODO: Fix that.